Fix the showing/hiding button depending on path rules

+++ app/frontend/svelte/apps/init-button.ts
@@ -1,6 +1,6 @@
 import Summer from '@/svelte/apps/summer/Summer.svelte';
 import type { SvelteComponent } from 'svelte';
-import { ArticleInitInfo, initButton, SettingsInfo } from '@/svelte/apps/summer/api';
+import { initButton, SettingsInfo } from '@/svelte/apps/summer/api';
 
 let buttonApp: SvelteComponent | undefined;
 let checkerInterval: number | undefined;
@@ -8,50 +8,61 @@ let timeoutId: number;
 
 const log = (message: string) => {
   console.log(`%c${message}`, 'background: #FFEFE7; padding: 2px 10px;');
-}
+};
 const destroyApp = () => {
   if (buttonApp && typeof buttonApp.$destroy == 'function') buttonApp.$destroy();
   buttonApp = undefined;
 };
 
-const installApp = (projectId: string, settings: SettingsInfo, article: ArticleInitInfo) => {
-  const appId = 'getsummer-' + projectId;
-  let appElement = document.getElementById(appId);
-  if (!appElement) {
-    appElement = document.createElement('div');
-    appElement.id = appId;
-    document.body.appendChild(appElement);
-  } else {
-    if (appElement.innerHTML !== '') appElement.innerHTML = '';
-    if (appElement.shadowRoot?.innerHTML) appElement.shadowRoot.innerHTML = '';
-  }
-  destroyApp();
-  buttonApp = new Summer({ target: appElement.attachShadow({ mode: "open" }), props: { projectId, settings, article } });
-  return appId;
+const createAppContainer = (appId: string) => {
+  const appElement = document.createElement('div');
+  appElement.id = appId;
+  document.body.appendChild(appElement);
+  return appElement;
+};
+
+const deleteAppContainerIfExists = (appId: string) => {
+  const appElement = document.getElementById(appId);
+  if (!appElement) return false;
+
+  if (appElement.innerHTML !== '') appElement.innerHTML = '';
+  if (appElement.shadowRoot?.innerHTML) appElement.shadowRoot.innerHTML = '';
+  appElement.remove();
+  return true;
+};
+
+const isPathValid = (pathname: string, paths: string[]) => {
+  const path = paths.find((path) => pathname.indexOf(path) === 0);
+  return typeof path !== 'undefined';
 };
 
 export const initApp = (project: { id: string; settings: object }, url: string, timeout = 100) => {
+  const appId = 'getsummer-' + project.id;
   const pathname = new URL(url).pathname;
-  const settings = project.settings as SettingsInfo
+  const settings = project.settings as SettingsInfo;
 
   clearInterval(checkerInterval);
   clearTimeout(timeoutId);
 
-  const path = settings.paths.find((path) => (pathname.indexOf(path) === 0))
-  // console.log('initApp', path, pathname);
-  if (typeof path === 'undefined') return;
+  deleteAppContainerIfExists(appId);
+  if (!isPathValid(pathname, settings.paths)) return;
 
   timeoutId = setTimeout(() => {
-
     initButton(project.id, url)
       .then((info) => {
-        const body = info.body
+        const body = info.body;
         if (!body) return log("GetSummer: Skipping page as there's no article");
         if ('code' in body) return console.error('GetSummer', info);
-        const install = () =>
-          installApp(project.id, settings, body.article);
-        // console.log(info);
-        const appId = install();
+
+        const install = () => {
+          if (deleteAppContainerIfExists(appId)) destroyApp();
+          buttonApp = new Summer({
+            target: createAppContainer(appId).attachShadow({ mode: 'open' }),
+            props: { projectId: project.id, settings, article: body.article },
+          });
+        };
+
+        install();
         checkerInterval = setInterval(() => {
           if (document.getElementById(appId) !== null) return;
           log('GetSummer: App destroyed, reinstalling');
